+++ src/main/java/com/senne/service/impl/OrderServiceImpl.java
@@ -56,7 +56,7 @@ public class OrderServiceImpl implements OrderService {
             createdOrder.setUser(user);
             createdOrder.setSellerId(sellerId);
             createdOrder.setTotalMrpPrice(totalOrderPrice);
-            createdOrder.setTotalSellingPrice(totalOrderPrice);
+            createdOrder.setTotalSellingPrice((long) totalOrderPrice);
             createdOrder.setTotalItem(totalItem);
             createdOrder.setShippingAddress(address);
